+++ saleor/graphql/translations/types.py
@@ -1,6 +1,7 @@
 import graphene
 from django.conf import settings
 
+from ...attribute import AttributeInputType
 from ...attribute import models as attribute_models
 from ...core.permissions import DiscountPermissions, ShippingPermissions
 from ...core.tracing import traced_resolver
@@ -74,7 +75,7 @@ class AttributeValueTranslatableContent(CountableDjangoObjectType):
     class Meta:
         model = attribute_models.AttributeValue
         interfaces = [graphene.relay.Node]
-        only_fields = BASIC_TRANSLATABLE_FIELDS
+        only_fields = BASIC_TRANSLATABLE_FIELDS + ["rich_text"]
 
     @staticmethod
     def resolve_attribute_value(root: attribute_models.AttributeValue, _info):
@@ -128,6 +129,11 @@ class ProductVariantTranslatableContent(CountableDjangoObjectType):
             "Will be removed in Saleor 4.0. " "Get model fields from the root level."
         ),
     )
+    attribute_values = graphene.List(
+        graphene.NonNull(AttributeValueTranslatableContent),
+        required=True,
+        description="List of product variant attribute values that can be translated.",
+    )
 
     class Meta:
         model = product_models.ProductVariant
@@ -137,6 +143,16 @@ class ProductVariantTranslatableContent(CountableDjangoObjectType):
     @staticmethod
     def resolve_product_variant(root: product_models.ProductVariant, info):
         return ChannelContext(node=root, channel_slug=None)
+
+    @staticmethod
+    def resolve_attribute_values(root: product_models.ProductVariant, _info):
+        translatable_values = []
+        for assignment in root.attributes.all():
+            attr = assignment.attribute
+            if attr.input_type in AttributeInputType.TRANSLATABLE_ATTRIBUTES:
+                value = assignment.values.first()
+                translatable_values.append(value)
+        return translatable_values
 
 
 class ProductTranslation(BaseTranslationType):
@@ -173,6 +189,11 @@ class ProductTranslatableContent(CountableDjangoObjectType):
             "Will be removed in Saleor 4.0. " "Get model fields from the root level."
         ),
     )
+    attribute_values = graphene.List(
+        graphene.NonNull(AttributeValueTranslatableContent),
+        required=True,
+        description="List of product attribute values that can be translated.",
+    )
 
     class Meta:
         model = product_models.Product
@@ -187,6 +208,16 @@ class ProductTranslatableContent(CountableDjangoObjectType):
     def resolve_description_json(root: product_models.Product, _info):
         description = root.description
         return description if description is not None else {}
+
+    @staticmethod
+    def resolve_attribute_values(root: product_models.Product, _info):
+        translatable_values = []
+        for assignment in root.attributes.all():
+            attr = assignment.attribute
+            if attr.input_type in AttributeInputType.TRANSLATABLE_ATTRIBUTES:
+                value = assignment.values.first()
+                translatable_values.append(value)
+        return translatable_values
 
 
 class CollectionTranslation(BaseTranslationType):
@@ -335,6 +366,11 @@ class PageTranslatableContent(CountableDjangoObjectType):
             "Will be removed in Saleor 4.0. " "Get model fields from the root level."
         ),
     )
+    attribute_values = graphene.List(
+        graphene.NonNull(AttributeValueTranslatableContent),
+        required=True,
+        description="List of page content attribute values that can be translated.",
+    )
 
     class Meta:
         model = page_models.Page
@@ -359,6 +395,16 @@ class PageTranslatableContent(CountableDjangoObjectType):
     def resolve_content_json(root: page_models.Page, _info):
         content = root.content
         return content if content is not None else {}
+
+    @staticmethod
+    def resolve_attribute_values(root: page_models.Page, _info):
+        translatable_values = []
+        for assignment in root.attributes.all():
+            attr = assignment.attribute
+            if attr.input_type in AttributeInputType.TRANSLATABLE_ATTRIBUTES:
+                value = assignment.values.first()
+                translatable_values.append(value)
+        return translatable_values
 
 
 class VoucherTranslation(BaseTranslationType):