+++ CHANGELOG.md
@@ -127,6 +127,7 @@ All notable, unreleased changes to this project will be documented in this file.
 - Fix failing account mutations for app - #7569 by @IKarbowiak
 - Modify order of auth middleware calls - #7572 by @tomaszszymanski129
 - Add app support for events - #7622 by @IKarbowiak
+- Add date & date time attributes - #7500 by @piotrgrundas
 
 ### Breaking
 - Multichannel MVP: Multicurrency - #6242 by @fowczarek @d-wysocki